Modifying scripts to use float32and clear memory after each iteration to prevent out of memory errors

+++ mobo2d/sumOfGaussians.py
@@ -2,7 +2,7 @@ import torch
 
 # Use GPU if possible
 tkwargs = {
-    "dtype": torch.double,
+    "dtype": torch.float,
     "device": torch.device("cuda" if torch.cuda.is_available() else "cpu"),
 }
